requirements.txt
fastapi
uvicorn[standard]
jinja2
mysql-connector-python
python-multipart
python-dotenv
